refactor: moving relic transforms into webgpu data transform
